Fix tdnf install in firstboot role
The tdnf command in the firstboot role is happening before any custom
repos are put in place. This means that if you happen to define a
private or internal mirror to use instead of a public one, the firstboot
role ignores this and uses the public repo anyways.

The logic in the setup role (meant to be used as a dependency) already
accounts for this, and needs to be used instead.