refactor role providers/vmware
A lot of distro-specific logic was in main.yml, while we already had a
couple of vmware-distro.yml files to be included. this patch adds a
vmware-redhat.yml file, and makes sure all vmware-distro files are
included by vmware.yml. This refactoring makes it much easier to see
what logic is running under what circumstances.